ffooo

+++ backend/loan-management-api/loan-backend/src/main/java/com/patrick/fintech/loan_backend/service/LoanService.java
@@ -2579,7 +2579,7 @@ public class LoanService {
                         }
 
                         loan.setRiskScore(
-                                        risk.getScore());
+                                        BigDecimal.valueOf(risk.getScore()));
 
                         loan.setRiskCategory(
                                         risk.getCategory());
